+++ lib/embedded_tweet_view.dart
@@ -49,6 +49,13 @@ class EmbeddedTweetView extends StatelessWidget {
   /// By default it is true
   final bool? enableVideoFullscreen;
 
+  /// Set video Control Bar background color
+  final Color? videoControlBarBgColor;
+
+  /// The placeholder is displayed underneath the Video before it is initialized
+  /// or played.
+  final Widget? videoPlaceholder;
+
   bool get _shouldShowReplies =>
       showRepliesCount && _tweetVM.repliesCount != null;
 
@@ -63,6 +70,8 @@ class EmbeddedTweetView extends StatelessWidget {
     this.showRepliesCount = false,
     this.autoPlayVideo,
     this.enableVideoFullscreen,
+    this.videoControlBarBgColor,
+    this.videoPlaceholder,
   }); //  TweetView(this.tweetVM);
 
   EmbeddedTweetView.fromTweetV1(
@@ -76,6 +85,8 @@ class EmbeddedTweetView extends StatelessWidget {
     this.showRepliesCount = false,
     this.autoPlayVideo,
     this.enableVideoFullscreen,
+    this.videoControlBarBgColor,
+    this.videoPlaceholder,
   }) : _tweetVM = TweetVM.fromApiModel(tweet, createdDateDisplayFormat);
 
   EmbeddedTweetView.fromTweetV2(
@@ -89,6 +100,8 @@ class EmbeddedTweetView extends StatelessWidget {
     this.showRepliesCount = false,
     this.autoPlayVideo,
     this.enableVideoFullscreen,
+    this.videoControlBarBgColor,
+    this.videoPlaceholder,
   }) : _tweetVM = TweetVM.fromApiV2Model(tweet, createdDateDisplayFormat);
 
   @override
@@ -216,6 +229,8 @@ class EmbeddedTweetView extends StatelessWidget {
               videoHighQuality: videoHighQuality,
               autoPlayVideo: autoPlayVideo,
               enableVideoFullscreen: enableVideoFullscreen,
+              videoControlBarBgColor: videoControlBarBgColor,
+              videoPlaceholder: videoPlaceholder,
             ),
           ),
           Container(