+++ lib/src/media_container.dart
@@ -22,6 +22,8 @@ class MediaContainer extends StatefulWidget {
     this.onTapImage,
     bool? autoPlayVideo = false,
     bool? enableVideoFullscreen = true,
+    this.videoControlBarBgColor,
+    this.videoPlaceholder,
   })  : autoPlayVideo = autoPlayVideo ?? false,
         enableVideoFullscreen = enableVideoFullscreen ?? true,
         super(key: key);
@@ -32,6 +34,8 @@ class MediaContainer extends StatefulWidget {
   final double? videoPlayerInitialVolume;
   final bool autoPlayVideo;
   final bool enableVideoFullscreen;
+  final Color? videoControlBarBgColor;
+  final Widget? videoPlaceholder;
 
   @override
   _MediaContainerState createState() => _MediaContainerState();
@@ -59,6 +63,8 @@ class _MediaContainerState extends State<MediaContainer>
         videoHighQuality: widget.videoHighQuality,
         autoPlay: widget.autoPlayVideo,
         enableFullscreen: widget.enableVideoFullscreen,
+        controlBarColor: widget.videoControlBarBgColor,
+        videoPlaceholder: widget.videoPlaceholder,
       );
     } else if (widget.tweetVM.getDisplayTweet().hasPhoto) {
       switch (widget.tweetVM.getDisplayTweet().allPhotos.length) {