+++ lib/src/tweet_video.dart
@@ -11,6 +11,8 @@ class TweetVideo extends StatefulWidget {
     this.autoPlay = false,
     this.enableFullscreen = true,
     this.videoHighQuality = true,
+    this.controlBarColor,
+    this.videoPlaceholder,
   }) : super(key: key);
 
   final TweetVM tweetVM;
@@ -18,6 +20,8 @@ class TweetVideo extends StatefulWidget {
   final bool autoPlay;
   final bool enableFullscreen;
   final bool? videoHighQuality;
+  final Color? controlBarColor;
+  final Widget? videoPlaceholder;
 
   @override
   _TweetVideoState createState() => _TweetVideoState();
@@ -32,18 +36,20 @@ class _TweetVideoState extends State<TweetVideo>
   void initState() {
     super.initState();
     betterPlayerConfiguration = BetterPlayerConfiguration(
-      placeholder: Center(
-        child: SizedBox(
-          height: 32,
-          width: 32,
-          child: CircularProgressIndicator(),
-        ),
-      ),
+      placeholder: widget.videoPlaceholder ??
+          Center(
+            child: SizedBox(
+              height: 32,
+              width: 32,
+              child: CircularProgressIndicator(),
+            ),
+          ),
       errorBuilder: (context, message) {
         return Text('Error while loading video :-(');
       },
       aspectRatio: widget.tweetVM.getDisplayTweet().videoAspectRatio!,
       controlsConfiguration: BetterPlayerControlsConfiguration(
+        controlBarColor: widget.controlBarColor ?? Colors.black87,
         enablePlaybackSpeed: false,
         enableSkips: false,
         enableMute: !widget.tweetVM.getDisplayTweet().hasGif,