+++ CHANGELOG.md
@@ -5,7 +5,17 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+
+## [4.0.4] - 13.09.2022
 ### Changed
+- Added videoPlaceholder for custom the video placeholder
+
+## [4.0.3] - 04.08.2022
+### Changed
+- Fixed url launcher deprecation - use canLaunchUrl and launchUrl instead of canLaunch and launch
+- Added videoControlBarBgColor for edit the video player control bar background
+
 
 ## [4.0.2] - 28.05.2022
 ### Changed